requirements.txt
numpy~=1.22.3
scipy~=1.3.3
Pillow~=9.0.1
matplotlib~=3.1.2